requirements.txt
# Core dependencies
huggingface-hub==0.16.4
sentence-transformers>=2.2.2
pinecone-client>=2.2.4
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API and web
fastapi>=0.100.0
uvicorn>=0.23.0

# Vector operations
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0 